moved barcode generator and barcode saver functions to BarcodeGeneratorAndSaver.kt added background color options for cards made card number a Long instead of Int

+++ app/src/main/java/com/example/cards_app/AppDataStore.kt
@@ -51,7 +51,7 @@ class AppDataStore(private val context: Context) {
         }
 
     }
-    suspend fun deleteCardByNumber(number: Int) {
+    suspend fun deleteCardByNumber(number: Long) {
         val currentCards = cardsFlow.first().toMutableList()//gets currentList<Card>
         val initialSize = currentCards.size //gets initial size
         currentCards.removeAll { it.number == number } //removes card with number
